% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_tp_error.R
\name{estimate_tp_error}
\alias{estimate_tp_error}
\title{Estimate error in trophic position}
\usage{
estimate_tp_error(
  sources,
  baselines,
  TP_b = 2,
  TDF_N,
  TDF_C = NULL,
  consumer_p
)
}
\arguments{
\item{sources}{A data frame with mean isotopic values of sources.
Must have a column named "d15N_mean" and, if using two baselines,
a column named "d13C_mean".}

\item{baselines}{A data frame with mean isotopic values of one or two baselines.
Must have a column named "d15N_mean" and, if using two baselines,
a column named "d13C_mean".}

\item{TP_b}{A numeric value. Trophic position of baseline(s).
Default to 2 (i.e., primary consumers).}

\item{TDF_N}{A numeric value. The TDF for \eqn{\delta^{15}\text{N}}.}

\item{TDF_C}{A numeric value. The TDF for \eqn{\delta^{13}\text{C}}. Default to NULL.
Only necessary for two baselines.}

\item{consumer_p}{A numeric vector summing to 1.
The source relative contributions to the consumer.}
}
\value{
A numeric value
}
\description{
Estimate error in trophic position
}
\examples{
library(dplyr)
library(tidyr)

# 3 sources
n_sources <- 3
sources <- data.frame(source = paste("source", 1:n_sources),
                      d13C_mean = rnorm(n_sources, 0, 5), 
                      d15N_mean = rnorm(n_sources, 0, 5))

# TDFs (values from Post 2002 for this example)
TDF_C <- 0.4
TDF_N <- 3.4

# 2 baselines
n_baselines <- 2

# Random number generation from the dirichlet distribution
# n: number of random vectors to generate
# alpha: vector of positive shape parameters
rdirichlet <- function (n, alpha) {
    x <- matrix(rgamma(length(alpha) * n, alpha), 
                ncol = length(alpha), 
                byrow = TRUE)
    return(x / rowSums(x))
}

# Generate baselines isotopic values:
# 1. randomly generate source proportions
# 2. multiply by source values and summarise
# 3. add TDF to set baselines at TP = 2
baselines <- rdirichlet(n = n_baselines, alpha = c(1, 1, 1)) |> 
  t() |> 
  as.data.frame() |> 
  setNames(paste("baseline", 1:n_baselines)) |>
  mutate(source = paste("source", 1:n_sources)) |> 
  pivot_longer(cols = starts_with("baseline"),
               names_to = "baseline",
               values_to = "p") |>
  left_join(sources) |>
  mutate(across(ends_with("_mean"), ~.x * p)) |>
  group_by(baseline) |>
  summarise(d13C_mean = sum(d13C_mean) + TDF_C, 
            d15N_mean = sum(d15N_mean) + TDF_N)

# Source relative contributions for 7 consumers
consumers_p <- list(c(0.8, 0.1, 0.1), # relies mostly on source 1
                    c(0.1, 0.8, 0.1), # relies mostly on source 2
                    c(0.1, 0.1, 0.8), # relies mostly on source 3
                    c(0.45, 0.45, 0.1), # relies mostly on source 1 and 2
                    c(0.45, 0.1, 0.45), # relies mostly on source 1 and 3
                    c(0.1, 0.45, 0.45), # relies mostly on source 2 and 3
                    c(1/3, 1/3, 1/3)) # relies equally on sources
names(consumers_p) <- paste("consumer", 1:7)

lapply(consumers_p,
       function(i) 
         estimate_tp_error(sources, baselines, TP_b = 2, TDF_N, TDF_C, i)
)

}
