% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_food_web.R
\name{rescale_dirichlet}
\alias{rescale_dirichlet}
\title{Rescale Dirichlet proportions to avoid values < 0.05}
\usage{
rescale_dirichlet(data, d)
}
\arguments{
\item{data}{A data frame with one column for each category and one row for each mixture.
Each row must sum up to 1}

\item{d}{Number of categories}
}
\value{
A data frame
}
\description{
Rescale Dirichlet proportions to avoid values < 0.05
}
