% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_food_web.R
\name{simulate_consumer_data}
\alias{simulate_consumer_data}
\title{Simulate consumer isotope data}
\usage{
simulate_consumer_data(
  n,
  sources,
  alpha,
  C = 10,
  TP_mean,
  TP_sd = 0.05,
  deltaC,
  deltaN,
  seed = NA
)
}
\arguments{
\item{n}{Number of observations to simulate}

\item{sources}{A data frame of source isotope data with three columns:
\itemize{
\item \code{name}: source names
\item \code{d13C_mean}: mean carbon stable isotope ratio
\item \code{d13C_sd}: standard deviation of carbon stable isotope ratio
\item \code{d15N_mean}: mean nitrogen stable isotope ratio
\item \code{d15N_sd}: standard deviation of nitrogen stable isotope ratio
}}

\item{alpha}{A numeric vector of known source proportions}

\item{C}{Total consumption rate (i.e. total number of source items eaten
by the consumer in the tissue turnover period). Default to 10}

\item{TP_mean}{Mean trophic position}

\item{TP_sd}{Standard deviation for trophic position. Default to 0.05}

\item{deltaC}{A numeric vector. Trophic discrimination factor for C}

\item{deltaN}{A numeric vector. Trophic discrimination factor for N}

\item{seed}{The seed passed to \code{\link[base:Random]{base::set.seed()}} to make simulations reproducible.
If \code{NA} (the default), \code{set.seed} will not be called}
}
\value{
A data frame with \code{n} rows and two columns, "d13C" and "d15N".
}
\description{
Simulate carbon and nitrogen stable isotope data for a consumer based on
known source proportions and trophic position
}
\examples{
sources <- data.frame(name = paste("source", 1:3),
                      d13C_mean = rnorm(3, 0, 5),
                      d13C_sd = runif(3, 0.5, 1.5),
                      d15N_mean = rnorm(3, 0, 5),
                      d15N_sd = runif(3, 0.5, 1.5))
alpha <- compositions::rDirichlet.acomp(1, c(1, 1, 1)) |> as.vector()
# TDFs from Post 2002
deltaC <- rnorm(59, 0.39, 1.69)
deltaN <- rnorm(107, 3.4, 0.96)
simulate_consumer_data(10, sources, alpha, C = 10, TP_mean = 3.5, TP_sd = 0.05, deltaC, deltaN)

}
