% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_food_web.R
\name{simulate_food_web}
\alias{simulate_food_web}
\title{Simulate a food web}
\usage{
simulate_food_web(
  n_sources = 3,
  source_C_range = c(-10, 10),
  source_N_range = c(-10, 10),
  source_C_mean = NULL,
  source_N_mean = NULL,
  source_C_sd = runif(n_sources, 0.5, 1.5),
  source_N_sd = runif(n_sources, 0.5, 1.5),
  n_obs_sources = 10,
  n_baselines = 1,
  n_obs_baselines = 10,
  n_families = 3,
  n_species = 10,
  n_obs_consumers = 10,
  sd_families = 1.5,
  sd_species = 0.5,
  TP_sd = 0.05,
  n_TDF_C = 56,
  mean_TDF_C = 0.39,
  sd_TDF_C = 1.3,
  n_TDF_N = 107,
  mean_TDF_N = 3.4,
  sd_TDF_N = 0.98,
  C_baselines = 20,
  C_consumers = 20,
  seed = NA
)
}
\arguments{
\item{n_sources}{Number of sources to simulate. Default to 3}

\item{source_C_range}{The range in source \eqn{\delta^{13}\text{C}}. A numeric vector of length 2 (minimum and maximum).
Ignored if \code{source_C_mean} is provided}

\item{source_N_range}{The range in source \eqn{\delta^{15}\text{N}}. A numeric vector of length 2 (minimum and maximum).
Ignored if \code{source_N_mean} is provided}

\item{source_C_mean}{Numeric vector of length \code{n_sources} with the mean \eqn{\delta^{13}\text{C}} for each source.
If NULL (default), mean \eqn{\delta^{13}\text{C}} values are sampled from equal (n_sources * 2) intervals along the given \code{source_C_range}}

\item{source_N_mean}{Numeric vector of length \code{n_sources} with the mean \eqn{\delta^{15}\text{N}} for each source.
If NULL (default), mean \eqn{\delta^{15}\text{N}} values are sampled from equal (n_sources * 2) intervals along the given \code{source_N_range}}

\item{source_C_sd}{Numeric vector of length \code{n_sources} with the \eqn{\delta^{13}\text{C}} standard deviation for each source}

\item{source_N_sd}{Numeric vector of length \code{n_sources} with the \eqn{\delta^{15}\text{N}} standard deviation for each source}

\item{n_obs_sources}{Numeric vector of length \code{n_sources} with the number of observations
for each source. Can be a single value if all source have the same number of observations.
Default to 10}

\item{n_baselines}{Number of baselines (i.e. primary consumers, TP = 2) to simulate.
Must be 1 (default) or 2}

\item{n_obs_baselines}{Numeric vector of length \code{n_baselines} with number of observations
for each baseline. Can be a single value if all baselines have the same number of observations.
Default to 10}

\item{n_families}{Number of consumer families to simulate. Default to 3}

\item{n_species}{Number of consumer species to simulate. Default to 10.
These are randomly split among the families}

\item{n_obs_consumers}{Numeric vector of length \code{n_species} with number of observations
for each consumer species. Can be a single value if all consumers have the same number of
observations. Default to 10}

\item{sd_families}{Variation in diet across families. Default to 1.5}

\item{sd_species}{Variation in diet across species. Default to 0.5}

\item{TP_sd}{Standard deviation of trophic position for baselines and consumer species. Default to 0.05}

\item{n_TDF_C}{Number of observations for carbon's trophic discrimination factor}

\item{mean_TDF_C}{Mean for carbon's trophic discrimination factor}

\item{sd_TDF_C}{Standard deviation for carbon's trophic discrimination factor}

\item{n_TDF_N}{Number of observations for nitrogen's trophic discrimination factor}

\item{mean_TDF_N}{Mean for nitrogen's trophic discrimination factor}

\item{sd_TDF_N}{Standard deviation for nitrogen's trophic discrimination factor}

\item{C_baselines, C_consumers}{Total consumption of baselines and consumers
(i.e. total number of source items eaten in the tissue turnover period). Default to 20}

\item{seed}{The seed passed to \code{\link[base:Random]{base::set.seed()}} to make simulations reproducible.
If \code{NA} (the default), \code{set.seed} will not be called}
}
\value{
A list with the following elements:
\itemize{
\item \code{isotopes}: data frame with isotope data for all observations
\item \code{sources_summary}: data frame with isotopic means and standard deviations for each source
\item \code{TP_families}: data frame with number of species, mean and
standard deviation of trophic position for each family
\item \code{TP_species}: data frame with trophic position for each species
\item \code{alpha_baselines}: data frame with source proportions for each baseline
\item \code{alpha_global}: data frame with global source proportions
\item \code{alpha_families}: data frame with source proportions for each family
\item \code{alpha_species}: data frame with source proportions for each species
\item \code{TDF}: list of length two including the simulated values of
C and N trophic discrimination factors
\item \code{TDF_summary}: data frame with mean and standard deviation of trophic
discrimination factors for each source
\item \code{isospace_plot}: ggplot object displaying the food web in isotopic space
}
}
\description{
Simulate carbon and nitrogen stable isotope data for a food web including
basal sources (i.e. primary producers), baselines (i.e. primary consumers),
and higher-level consumers, organised into species and families.
}
\examples{
simulate_food_web()

}
